Start work on the RyR constitutively bound to CaM
If RyR has really indeed high affinity to CaM and apoCaM, in the cytosol every RyR will be bound to 4 CaM. I have not yet found an in silico experiment that will require all states (maybe AD, but this will be done later). For the scope of this project RyRCam and RyR not binding to CaM should be enough.